backend_api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dateutil==2.8.2
google-generativeai>=0.8.0
langgraph>=0.6.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
rank-bm25==0.2.2
python-dotenv==1.0.0
